move ES5 classes to ES6
