+++ lib/DirectoryWatcher.js
@@ -9,324 +9,12 @@ const path = require('path');
 const async = require('async');
 const chokidar = require('chokidar');
 const fs = require('graceful-fs');
-const watcherManager = require('./watcherManager');
+const Watcher = require('./Watcher');
+const WatcherManager = require('./WatcherManager');
+
+const manager = new WatcherManager();
 
 let FS_ACCURACY = 10000;
-
-
-function withoutCase(str) {
-  return str.toLowerCase();
-}
-
-
-function Watcher(directoryWatcher, filePath, startTime) {
-  EventEmitter.call(this);
-  this.directoryWatcher = directoryWatcher;
-  this.path = filePath;
-  this.startTime = startTime && +startTime;
-  this.data = 0;
-}
-
-Watcher.prototype = Object.create(EventEmitter.prototype);
-Watcher.prototype.constructor = Watcher;
-
-Watcher.prototype.checkStartTime = function checkStartTime(mtime, initial) {
-  if (typeof this.startTime !== 'number') return !initial;
-  const startTime = this.startTime;
-  return startTime <= mtime;
-};
-
-Watcher.prototype.close = function close() {
-  this.emit('closed');
-};
-
-
-function DirectoryWatcher(directoryPath, options) {
-  EventEmitter.call(this);
-  this.options = options;
-  this.path = directoryPath;
-  this.files = Object.create(null);
-  this.directories = Object.create(null);
-  this.watcher = chokidar.watch(directoryPath, {
-    ignoreInitial: true,
-    persistent: true,
-    followSymlinks: false,
-    depth: 0,
-    atomic: false,
-    alwaysStat: true,
-    ignorePermissionErrors: true,
-    ignored: options.ignored,
-    usePolling: options.poll ? true : undefined, // eslint-disable-line no-undefined
-    interval: typeof options.poll === 'number' ? options.poll : undefined, // eslint-disable-line no-undefined
-    disableGlobbing: true
-  });
-  this.watcher.on('add', this.onFileAdded.bind(this));
-  this.watcher.on('addDir', this.onDirectoryAdded.bind(this));
-  this.watcher.on('change', this.onChange.bind(this));
-  this.watcher.on('unlink', this.onFileUnlinked.bind(this));
-  this.watcher.on('unlinkDir', this.onDirectoryUnlinked.bind(this));
-  this.watcher.on('error', this.onWatcherError.bind(this));
-  this.initialScan = true;
-  this.nestedWatching = false;
-  this.initialScanRemoved = [];
-  this.doInitialScan();
-  this.watchers = Object.create(null);
-  this.refs = 0;
-}
-module.exports = DirectoryWatcher;
-
-DirectoryWatcher.prototype = Object.create(EventEmitter.prototype);
-DirectoryWatcher.prototype.constructor = DirectoryWatcher;
-
-DirectoryWatcher.prototype.setFileTime = function setFileTime(filePath, mtime, initial, type) {
-  const now = Date.now();
-  const old = this.files[filePath];
-
-  this.files[filePath] = [initial ? Math.min(now, mtime) : now, mtime];
-
-  // we add the fs accurency to reach the maximum possible mtime
-  if (mtime) { mtime += FS_ACCURACY; }
-
-  if (!old) {
-    if (mtime) {
-      if (this.watchers[withoutCase(filePath)]) {
-        this.watchers[withoutCase(filePath)].forEach((w) => {
-          if (!initial || w.checkStartTime(mtime, initial)) {
-            w.emit('change', mtime, initial ? 'initial' : type);
-          }
-        });
-      }
-    }
-  } else if (!initial && mtime && type !== 'add') {
-    if (this.watchers[withoutCase(filePath)]) {
-      this.watchers[withoutCase(filePath)].forEach((w) => {
-        w.emit('change', mtime, type);
-      });
-    }
-  } else if (!initial && !mtime) {
-    if (this.watchers[withoutCase(filePath)]) {
-      this.watchers[withoutCase(filePath)].forEach((w) => {
-        w.emit('remove', type);
-      });
-    }
-  }
-  if (this.watchers[withoutCase(this.path)]) {
-    this.watchers[withoutCase(this.path)].forEach((w) => {
-      if (!initial || w.checkStartTime(mtime, initial)) {
-        w.emit('change', filePath, mtime, initial ? 'initial' : type);
-      }
-    });
-  }
-};
-
-DirectoryWatcher.prototype.setDirectory = function setDirectory(directoryPath, exist, initial, type) {
-  if (directoryPath === this.path) {
-    if (!initial && this.watchers[withoutCase(this.path)]) {
-      this.watchers[withoutCase(this.path)].forEach((w) => {
-        w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
-      });
-    }
-  } else {
-    const old = this.directories[directoryPath];
-    if (!old) {
-      if (exist) {
-        if (this.nestedWatching) {
-          this.createNestedWatcher(directoryPath);
-        } else {
-          this.directories[directoryPath] = true;
-        }
-        if (!initial && this.watchers[withoutCase(this.path)]) {
-          this.watchers[withoutCase(this.path)].forEach((w) => {
-            w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
-          });
-        }
-      }
-    } else if (!exist) {
-      if (this.nestedWatching) { this.directories[directoryPath].close(); }
-      delete this.directories[directoryPath];
-      if (!initial && this.watchers[withoutCase(this.path)]) {
-        this.watchers[withoutCase(this.path)].forEach((w) => {
-          w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
-        });
-      }
-    }
-  }
-};
-
-DirectoryWatcher.prototype.createNestedWatcher = function createNestedWatcher(directoryPath) {
-  this.directories[directoryPath] = watcherManager.watchDirectory(directoryPath, this.options, 1);
-  this.directories[directoryPath].on('change', (filePath, mtime, type) => {
-    if (this.watchers[withoutCase(this.path)]) {
-      this.watchers[withoutCase(this.path)].forEach((w) => {
-        if (w.checkStartTime(mtime, false)) {
-          w.emit('change', filePath, mtime, type);
-        }
-      });
-    }
-  });
-};
-
-DirectoryWatcher.prototype.setNestedWatching = function setNestedWatching(flag) {
-  if (this.nestedWatching !== !!flag) {
-    this.nestedWatching = !!flag;
-    if (this.nestedWatching) {
-      Object.keys(this.directories).forEach(function fe(directory) {
-        this.createNestedWatcher(directory);
-      }, this);
-    } else {
-      Object.keys(this.directories).forEach(function fe(directory) {
-        this.directories[directory].close();
-        this.directories[directory] = true;
-      }, this);
-    }
-  }
-};
-
-DirectoryWatcher.prototype.watch = function watch(filePath, startTime) {
-  this.watchers[withoutCase(filePath)] = this.watchers[withoutCase(filePath)] || [];
-  this.refs += 1;
-  const watcher = new Watcher(this, filePath, startTime);
-  watcher.on('closed', () => {
-    const idx = this.watchers[withoutCase(filePath)].indexOf(watcher);
-    this.watchers[withoutCase(filePath)].splice(idx, 1);
-    if (this.watchers[withoutCase(filePath)].length === 0) {
-      delete this.watchers[withoutCase(filePath)];
-      if (this.path === filePath) { this.setNestedWatching(false); }
-    }
-    if (--this.refs <= 0) { this.close(); } // eslint-disable-line no-plusplus
-  });
-  this.watchers[withoutCase(filePath)].push(watcher);
-  let data;
-  if (filePath === this.path) {
-    this.setNestedWatching(true);
-    data = false;
-    Object.keys(this.files).forEach(function fe(file) {
-      const d = this.files[file];
-      if (!data) { data = d; } else { data = [Math.max(data[0], d[0]), Math.max(data[1], d[1])]; }
-    }, this);
-  } else {
-    data = this.files[filePath];
-  }
-  process.nextTick(() => {
-    if (data) {
-      const ts = data[0] === data[1] ? data[0] + FS_ACCURACY : data[0];
-      if (ts >= startTime) { watcher.emit('change', data[1]); }
-    } else if (this.initialScan && this.initialScanRemoved.indexOf(filePath) >= 0) {
-      watcher.emit('remove');
-    }
-  });
-  return watcher;
-};
-
-DirectoryWatcher.prototype.onFileAdded = function onFileAdded(filePath, stat) {
-  if (filePath.indexOf(this.path) !== 0) return;
-  if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
-
-  this.setFileTime(filePath, +stat.mtime, false, 'add');
-};
-
-DirectoryWatcher.prototype.onDirectoryAdded = function onDirectoryAdded(directoryPath /* , stat */) {
-  if (directoryPath.indexOf(this.path) !== 0) return;
-  if (/[\\/]/.test(directoryPath.substr(this.path.length + 1))) return;
-  this.setDirectory(directoryPath, true, false, 'add');
-};
-
-DirectoryWatcher.prototype.onChange = function onChange(filePath, stat) {
-  if (filePath.indexOf(this.path) !== 0) return;
-  if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
-  const mtime = +stat.mtime;
-  ensureFsAccuracy(mtime);
-  this.setFileTime(filePath, mtime, false, 'change');
-};
-
-DirectoryWatcher.prototype.onFileUnlinked = function onFileUnlinked(filePath) {
-  if (filePath.indexOf(this.path) !== 0) return;
-  if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
-  this.setFileTime(filePath, null, false, 'unlink');
-  if (this.initialScan) {
-    this.initialScanRemoved.push(filePath);
-  }
-};
-
-DirectoryWatcher.prototype.onDirectoryUnlinked = function onDirectoryUnlinked(directoryPath) {
-  if (directoryPath.indexOf(this.path) !== 0) return;
-  if (/[\\/]/.test(directoryPath.substr(this.path.length + 1))) return;
-  this.setDirectory(directoryPath, false, false, 'unlink');
-  if (this.initialScan) {
-    this.initialScanRemoved.push(directoryPath);
-  }
-};
-
-DirectoryWatcher.prototype.onWatcherError = function onWatcherError(/* err */) {
-};
-
-DirectoryWatcher.prototype.doInitialScan = function doInitialScan() {
-  fs.readdir(this.path, (err, items) => {
-    if (err) {
-      this.initialScan = false;
-      return;
-    }
-    async.forEach(items, (item, callback) => {
-      const itemPath = path.join(this.path, item);
-      fs.stat(itemPath, (err2, stat) => {
-        if (!this.initialScan) return;
-        if (err2) {
-          callback();
-          return;
-        }
-        if (stat.isFile()) {
-          if (!this.files[itemPath]) { this.setFileTime(itemPath, +stat.mtime, true); }
-        } else if (stat.isDirectory()) {
-          if (!this.directories[itemPath]) { this.setDirectory(itemPath, true, true); }
-        }
-        callback();
-      });
-    }, () => {
-      this.initialScan = false;
-      this.initialScanRemoved = null;
-    });
-  });
-};
-
-DirectoryWatcher.prototype.getTimes = function getTimes() {
-  const obj = Object.create(null);
-  let selfTime = 0;
-  Object.keys(this.files).forEach(function fe(file) {
-    const data = this.files[file];
-    let time;
-    if (data[1]) {
-      time = Math.max(data[0], data[1] + FS_ACCURACY);
-    } else {
-      time = data[0];
-    }
-    obj[file] = time;
-    if (time > selfTime) { selfTime = time; }
-  }, this);
-  if (this.nestedWatching) {
-    Object.keys(this.directories).forEach(function fe(dir) {
-      const w = this.directories[dir];
-      const times = w.directoryWatcher.getTimes();
-      Object.keys(times).forEach((file) => {
-        const time = times[file];
-        obj[file] = time;
-        if (time > selfTime) { selfTime = time; }
-      });
-    }, this);
-    obj[this.path] = selfTime;
-  }
-  return obj;
-};
-
-DirectoryWatcher.prototype.close = function close() {
-  this.initialScan = false;
-  this.watcher.close();
-  if (this.nestedWatching) {
-    Object.keys(this.directories).forEach(function fe(dir) {
-      this.directories[dir].close();
-    }, this);
-  }
-  this.emit('closed');
-};
 
 function ensureFsAccuracy(mtime) {
   if (!mtime) return;
@@ -342,3 +30,294 @@ function ensureFsAccuracy(mtime) {
     FS_ACCURACY = 2000;
   }
 }
+
+function withoutCase(str) {
+  return str.toLowerCase();
+}
+
+module.exports = class DirectoryWatcher extends EventEmitter {
+  constructor(directoryPath, options) {
+    super();
+
+    this.options = options;
+    this.path = directoryPath;
+    this.files = Object.create(null);
+    this.directories = Object.create(null);
+    this.watcher = chokidar.watch(directoryPath, {
+      ignoreInitial: true,
+      persistent: true,
+      followSymlinks: false,
+      depth: 0,
+      atomic: false,
+      alwaysStat: true,
+      ignorePermissionErrors: true,
+      ignored: options.ignored,
+      usePolling: options.poll ? true : undefined, // eslint-disable-line no-undefined
+      interval: typeof options.poll === 'number' ? options.poll : undefined, // eslint-disable-line no-undefined
+      disableGlobbing: true
+    });
+    this.watcher.on('add', this.onFileAdded.bind(this));
+    this.watcher.on('addDir', this.onDirectoryAdded.bind(this));
+    this.watcher.on('change', this.onChange.bind(this));
+    this.watcher.on('unlink', this.onFileUnlinked.bind(this));
+    this.watcher.on('unlinkDir', this.onDirectoryUnlinked.bind(this));
+    this.watcher.on('error', this.onWatcherError.bind(this));
+    this.initialScan = true;
+    this.nestedWatching = false;
+    this.initialScanRemoved = [];
+    this.doInitialScan();
+    this.watchers = Object.create(null);
+    this.refs = 0;
+  }
+
+  setFileTime(filePath, mtime, initial, type) {
+    const now = Date.now();
+    const old = this.files[filePath];
+
+    this.files[filePath] = [initial ? Math.min(now, mtime) : now, mtime];
+
+    // we add the fs accurency to reach the maximum possible mtime
+    if (mtime) { mtime += FS_ACCURACY; }
+
+    if (!old) {
+      if (mtime) {
+        if (this.watchers[withoutCase(filePath)]) {
+          this.watchers[withoutCase(filePath)].forEach((w) => {
+            if (!initial || w.checkStartTime(mtime, initial)) {
+              w.emit('change', mtime, initial ? 'initial' : type);
+            }
+          });
+        }
+      }
+    } else if (!initial && mtime && type !== 'add') {
+      if (this.watchers[withoutCase(filePath)]) {
+        this.watchers[withoutCase(filePath)].forEach((w) => {
+          w.emit('change', mtime, type);
+        });
+      }
+    } else if (!initial && !mtime) {
+      if (this.watchers[withoutCase(filePath)]) {
+        this.watchers[withoutCase(filePath)].forEach((w) => {
+          w.emit('remove', type);
+        });
+      }
+    }
+    if (this.watchers[withoutCase(this.path)]) {
+      this.watchers[withoutCase(this.path)].forEach((w) => {
+        if (!initial || w.checkStartTime(mtime, initial)) {
+          w.emit('change', filePath, mtime, initial ? 'initial' : type);
+        }
+      });
+    }
+  }
+
+  setDirectory(directoryPath, exist, initial, type) {
+    if (directoryPath === this.path) {
+      if (!initial && this.watchers[withoutCase(this.path)]) {
+        this.watchers[withoutCase(this.path)].forEach((w) => {
+          w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
+        });
+      }
+    } else {
+      const old = this.directories[directoryPath];
+      if (!old) {
+        if (exist) {
+          if (this.nestedWatching) {
+            this.createNestedWatcher(directoryPath);
+          } else {
+            this.directories[directoryPath] = true;
+          }
+          if (!initial && this.watchers[withoutCase(this.path)]) {
+            this.watchers[withoutCase(this.path)].forEach((w) => {
+              w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
+            });
+          }
+        }
+      } else if (!exist) {
+        if (this.nestedWatching) { this.directories[directoryPath].close(); }
+        delete this.directories[directoryPath];
+        if (!initial && this.watchers[withoutCase(this.path)]) {
+          this.watchers[withoutCase(this.path)].forEach((w) => {
+            w.emit('change', directoryPath, w.data, initial ? 'initial' : type);
+          });
+        }
+      }
+    }
+  }
+
+  createNestedWatcher(directoryPath) {
+    this.directories[directoryPath] = manager.watchDirectory(directoryPath, this.options, 1);
+    this.directories[directoryPath].on('change', (filePath, mtime, type) => {
+      if (this.watchers[withoutCase(this.path)]) {
+        this.watchers[withoutCase(this.path)].forEach((w) => {
+          if (w.checkStartTime(mtime, false)) {
+            w.emit('change', filePath, mtime, type);
+          }
+        });
+      }
+    });
+  }
+
+  setNestedWatching(flag) {
+    if (this.nestedWatching !== !!flag) {
+      this.nestedWatching = !!flag;
+      if (this.nestedWatching) {
+        Object.keys(this.directories).forEach(function fe(directory) {
+          this.createNestedWatcher(directory);
+        }, this);
+      } else {
+        Object.keys(this.directories).forEach(function fe(directory) {
+          this.directories[directory].close();
+          this.directories[directory] = true;
+        }, this);
+      }
+    }
+  }
+
+  watch(filePath, startTime) {
+    this.watchers[withoutCase(filePath)] = this.watchers[withoutCase(filePath)] || [];
+    this.refs += 1;
+    const watcher = new Watcher(this, filePath, startTime);
+    watcher.on('closed', () => {
+      const idx = this.watchers[withoutCase(filePath)].indexOf(watcher);
+      this.watchers[withoutCase(filePath)].splice(idx, 1);
+      if (this.watchers[withoutCase(filePath)].length === 0) {
+        delete this.watchers[withoutCase(filePath)];
+        if (this.path === filePath) { this.setNestedWatching(false); }
+      }
+      if (--this.refs <= 0) { this.close(); } // eslint-disable-line no-plusplus
+    });
+    this.watchers[withoutCase(filePath)].push(watcher);
+    let data;
+    if (filePath === this.path) {
+      this.setNestedWatching(true);
+      data = false;
+      Object.keys(this.files).forEach(function fe(file) {
+        const d = this.files[file];
+        if (!data) { data = d; } else { data = [Math.max(data[0], d[0]), Math.max(data[1], d[1])]; }
+      }, this);
+    } else {
+      data = this.files[filePath];
+    }
+    process.nextTick(() => {
+      if (data) {
+        const ts = data[0] === data[1] ? data[0] + FS_ACCURACY : data[0];
+        if (ts >= startTime) { watcher.emit('change', data[1]); }
+      } else if (this.initialScan && this.initialScanRemoved.indexOf(filePath) >= 0) {
+        watcher.emit('remove');
+      }
+    });
+    return watcher;
+  }
+
+  onFileAdded(filePath, stat) {
+    if (filePath.indexOf(this.path) !== 0) return;
+    if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
+
+    this.setFileTime(filePath, +stat.mtime, false, 'add');
+  }
+
+  onDirectoryAdded(directoryPath /* , stat */) {
+    if (directoryPath.indexOf(this.path) !== 0) return;
+    if (/[\\/]/.test(directoryPath.substr(this.path.length + 1))) return;
+    this.setDirectory(directoryPath, true, false, 'add');
+  }
+
+  onChange(filePath, stat) {
+    if (filePath.indexOf(this.path) !== 0) return;
+    if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
+    const mtime = +stat.mtime;
+    ensureFsAccuracy(mtime);
+    this.setFileTime(filePath, mtime, false, 'change');
+  }
+
+  onFileUnlinked(filePath) {
+    if (filePath.indexOf(this.path) !== 0) return;
+    if (/[\\/]/.test(filePath.substr(this.path.length + 1))) return;
+    this.setFileTime(filePath, null, false, 'unlink');
+    if (this.initialScan) {
+      this.initialScanRemoved.push(filePath);
+    }
+  }
+
+  onDirectoryUnlinked(directoryPath) {
+    if (directoryPath.indexOf(this.path) !== 0) return;
+    if (/[\\/]/.test(directoryPath.substr(this.path.length + 1))) return;
+    this.setDirectory(directoryPath, false, false, 'unlink');
+    if (this.initialScan) {
+      this.initialScanRemoved.push(directoryPath);
+    }
+  }
+
+  onWatcherError(/* err */) { // eslint-disable-line
+
+  }
+
+  doInitialScan() {
+    fs.readdir(this.path, (err, items) => {
+      if (err) {
+        this.initialScan = false;
+        return;
+      }
+      async.forEach(items, (item, callback) => {
+        const itemPath = path.join(this.path, item);
+        fs.stat(itemPath, (err2, stat) => {
+          if (!this.initialScan) return;
+          if (err2) {
+            callback();
+            return;
+          }
+          if (stat.isFile()) {
+            if (!this.files[itemPath]) { this.setFileTime(itemPath, +stat.mtime, true); }
+          } else if (stat.isDirectory()) {
+            if (!this.directories[itemPath]) { this.setDirectory(itemPath, true, true); }
+          }
+          callback();
+        });
+      }, () => {
+        this.initialScan = false;
+        this.initialScanRemoved = null;
+      });
+    });
+  }
+
+  getTimes() {
+    const obj = Object.create(null);
+    let selfTime = 0;
+    Object.keys(this.files).forEach(function fe(file) {
+      const data = this.files[file];
+      let time;
+      if (data[1]) {
+        time = Math.max(data[0], data[1] + FS_ACCURACY);
+      } else {
+        time = data[0];
+      }
+      obj[file] = time;
+      if (time > selfTime) { selfTime = time; }
+    }, this);
+    if (this.nestedWatching) {
+      Object.keys(this.directories).forEach(function fe(dir) {
+        const w = this.directories[dir];
+        const times = w.directoryWatcher.getTimes();
+        Object.keys(times).forEach((file) => {
+          const time = times[file];
+          obj[file] = time;
+          if (time > selfTime) { selfTime = time; }
+        });
+      }, this);
+      obj[this.path] = selfTime;
+    }
+    return obj;
+  }
+
+  close() {
+    this.initialScan = false;
+    this.watcher.close();
+    if (this.nestedWatching) {
+      Object.keys(this.directories).forEach(function fe(dir) {
+        this.directories[dir].close();
+      }, this);
+    }
+    this.emit('closed');
+  }
+};
